requirements.txt
playwright
markdownify
openai
python-dotenv